client_dimension_pipeline/client_dimension_sync_full/requirements.txt
google-cloud-bigquery>=2.34.4
google-cloud-secretmanager>=2.16.1
snowflake-connector-python>=3.0.1
functions-framework>=3.3.0 